@@ -8,7 +8,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.10.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
@@ -17,14 +17,14 @@ dependencies:
     sdk: flutter
 
   intl: ^0.17.0
-  url_launcher: ^6.0.3
-  provider: ^5.0.0
-  get_it: ^6.1.1
+  url_launcher: ^6.0.17
+  provider: ^6.0.1
+  get_it: ^7.2.0
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  mockito: ^4.1.1
-  lint: ^1.2.0
+  mockito: ^5.0.17
+  lint: ^1.8.1
 
 
 # For information on the generic Dart part of this file, see the
