@@ -11,15 +11,15 @@ class VolunteerData {
   final String role;
 
   /// Url of the website of the association
-  final String url;
+  final String? url;
 
   /// Details about what you did in the association
   final String details;
 
   VolunteerData(
-      {@required this.logoPath,
-      @required this.semanticsLabel,
-      @required this.role,
-      @required this.details,
+      {required this.logoPath,
+      required this.semanticsLabel,
+      required this.role,
+      required this.details,
       this.url});
 }
