@@ -2,10 +2,10 @@
 import 'package:flutter/material.dart';
 
 class LocaleProvider with ChangeNotifier {
-  Locale _locale;
-  Locale get locale => _locale;
+  Locale? _locale;
+  Locale? get locale => _locale;
 
-  void changeLocale(Locale locale) {
+  void changeLocale(Locale? locale) {
     _locale = locale;
 
     notifyListeners();
