@@ -47,7 +47,7 @@ class MyApp extends StatelessWidget {
 class Portfolio extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
-    AppTheme.instance.init(MediaQuery.of(context));
+    AppTheme.instance!.init(MediaQuery.of(context));
 
     return Scaffold(
       body: SafeArea(
