@@ -15,17 +15,17 @@ class ProjectsSection extends StatelessWidget {
       width: size.width,
       color: AppTheme.white,
       child: Padding(
-        padding: AppTheme.instance.sectionSpacing,
+        padding: AppTheme.instance!.sectionSpacing,
         child: Column(
           mainAxisAlignment: MainAxisAlignment.start,
           crossAxisAlignment: CrossAxisAlignment.start,
           children: [
-            Text(AppIntl.of(context).projects_title, style: AppTheme.instance.theme.textTheme.headline2),
-            SizedBox(height: AppTheme.instance.largeVerticalSpacing),
+            Text(AppIntl.of(context)!.projects_title, style: AppTheme.instance!.theme.textTheme.headline2),
+            SizedBox(height: AppTheme.instance!.largeVerticalSpacing),
             Center(
               child: Wrap(
                 spacing: 50,
-                children: _buildProjectCards(projectsData(AppIntl.of(context))),
+                children: _buildProjectCards(projectsData(AppIntl.of(context)!)),
               ),
             )
           ],
